# Performance optimization of SendPacketAsync