We revert to `PipeWrtier.WriteAsync` when payload of `MqttPacketBuffer` is emtpy.
| Method |     Mean |   Error |  StdDev | Ratio | Rank |   Gen0 | Allocated | Alloc Ratio |
|------- |---------:|--------:|--------:|------:|-----:|-------:|----------:|------------:|
| Before | 270.3 ns | 5.45 ns | 5.35 ns |  1.00 |    2 | 0.2241 |     352 B |        1.00 |
|  After | 178.1 ns | 3.54 ns | 3.93 ns |  0.66 |    1 | 0.1631 |     256 B |        0.73 |